requirements.txt
psycopg2-binary>=2.9.0
pydantic>=2.0.0